Import cleanup, more forced pure seperation
