+++ src/commands/math.ts
@@ -6,7 +6,6 @@ import { MQNode } from "../services/keystroke";
 import { CursorOptions, NodeRef, MathspeakOptions, LatexCmdsSingleCharBuilder } from "../shared_types";
 import { Anticursor, Cursor, MQSelection } from "../cursor"
 import { API, APIClasses, RootBlockMixin } from "../publicapi"
-import { Fragment, Ends, NodeBase } from "../tree";
 import { Direction, pray, L, R, noop } from "../utils"
 import { Parser, } from "../services/parser.util";
 import { domFrag } from "../domFragment";
@@ -18,6 +17,7 @@ import { Digit, Letter } from "./math/basicSymbols";
 import { Options } from "src/options";
 import { CharCmdsAny, InnerFields, JoinMethod, LatexCmdsAny } from "src/pure_types";
 import { LatexCmds, CharCmds } from "./list";
+import { Ends, Fragment, NodeBase } from "src/node";
 
 /**
  * Math tree node base class.
