+++ src/commands/math/basicSymbols.ts
@@ -1,6 +1,7 @@
 import { Cursor } from "src/cursor";
 import { h } from "src/dom";
 import { domFrag } from "src/domFragment";
+import { Fragment } from "src/node";
 import { Options } from "src/options";
 import { optionProcessors } from "src/publicapi";
 import { AutoDict, InequalityData, LatexCmdsAny } from "src/pure_types";
@@ -8,7 +9,7 @@ import { MQNode } from "src/services/keystroke";
 import { latexMathParser, TempSingleCharNode } from "src/services/latex";
 import { Parser } from "src/services/parser.util";
 import { CursorOptions, NodeRef, MathspeakOptions } from "src/shared_types";
-import { Fragment, isMQNodeClass } from "src/tree";
+import { isMQNodeClass } from "src/tree";
 import { U_NO_BREAK_SPACE } from "src/unicode";
 import { Direction, L, R } from "src/utils";
 import { LatexCmds, CharCmds } from "../list";
