+++ src/dom.ts
@@ -1,7 +1,7 @@
 import { pray } from "./utils"
 import { domFrag } from "./domFragment"
 import { MathBlock } from "./commands/math";
-import { NodeBase } from "./tree";
+import { NodeBase } from "./node";
 
 export const urlParams = new URLSearchParams(window.location.search);
