+++ src/pure_types.ts
@@ -1,5 +1,7 @@
 export type HandlerOptions = any;
 
+export type KIND_OF_MQ = 'StaticMath' | 'MathField' | 'InnerMathField' | 'TextField';
+
 export type ControllerEvent =
   | 'move'
   | 'upDown'
@@ -38,3 +40,19 @@ export type EmbedOptionsData = any;
 export type MQ = any;
 export type LatexCmdsAny = any;
 export type CharCmdsAny = any;
+
+export type HandlerName = any;
+export type JQ_KeyboardEvent = KeyboardEvent & {
+  originalEvent?: KeyboardEvent;
+};
+export type RootBlockMixinInput = any;
+export type ControllerData = any;
+
+export type JQSelector =
+  | JQuery
+  | HTMLElement
+  | null
+  | Window
+  | NodeListOf<ChildNode>
+  | HTMLElement[]
+  | EventTarget;
